Refactor game action mechanism to be more type-friendly
Most game actions just return an action now and in case of exceptional
circumstances exceptions are used to communicate the problem to
user_controller which displays the message. That way the actions
themselves can be just concerned with the happy path.

Refactor Visionary and Hoarder (Old HasVision and HasInventory) to
directly baseclass Creature which had seemingly always been an option to
do. Solves a lot of typing problems that was caused by the complicated
way of doing the mixins.

Added typing to controllers.* and game_actions.py at least

Add from __future__ import annotations to all files