Refactor dummy input
Dummy input is now injected into to the game more or less and is mostly
a separate mechanism from the game itself.

Add options to the dummy system to use the regular window systems
instead of MockWrapper to display what's happening.

Add options to continue the game after an integration test